Calculates some basic aggregate stats for a specified player

Calculates some basic aggregate stats for a specified player

Usage

stats_for_player(df, name, game = NULL)

Arguments

df

A data frame with data from the rocket league stats api

name

A string specifying the gamer-tag of the desired player

game

Optional. A string specifying the guid of the desired match to analyze. Otherwise aggregates over all matches.

Value

Some aggregate statistics for the specified player

Examples

stats_for_player(TDMZ443games, name = "TDMZ443")

  

Stats for TDMZ443
    

Number_of_Ball_Touches
      	Number_Of_Goals
      	Number_Of_Saves
      	Average_Speed
      	Average_Boost
      	Percentage_Of_Time_Supersonic
    

68
	1
	4
	51.66
	54.81
	16.8%

stats_for_player(TDMZ443games, name = "TDMZ443", "D96D303C11F15617E1F59A8137901B26")

  

Stats for TDMZ443
    

Number_of_Ball_Touches
      	Number_Of_Goals
      	Number_Of_Saves
      	Average_Speed
      	Average_Boost
      	Percentage_Of_Time_Supersonic
    

68
	1
	1
	50.73
	54.40
	16.3%